refactor: merge interface with class since it's not reused
(cherry picked from commit 042e83b41ea88f130e5baf379c0d8af0faf275e6)
