+++ zeebe/broker/src/test/java/io/camunda/zeebe/broker/test/TestBrokerClientFactory.java
@@ -26,7 +26,7 @@ public final class TestBrokerClientFactory {
     final var topologyManager =
         new BrokerTopologyManagerImpl(
             () -> atomixCluster.getMembershipService().getMembers(),
-            BrokerClientTopologyMetrics.of(meterRegistry));
+            new BrokerClientTopologyMetrics(meterRegistry));
     actorScheduler.submitActor(topologyManager).join();
     atomixCluster.getMembershipService().addListener(topologyManager);
 
@@ -37,7 +37,7 @@ public final class TestBrokerClientFactory {
             atomixCluster.getEventService(),
             actorScheduler,
             topologyManager,
-            BrokerClientRequestMetrics.of(meterRegistry));
+            new BrokerClientRequestMetrics(meterRegistry));
     client.start().forEach(ActorFuture::join);
     return client;
   }
