+++ zeebe/gateway-grpc/src/test/java/io/camunda/zeebe/gateway/api/UnavailableBrokersTest.java
@@ -29,7 +29,10 @@ import io.camunda.zeebe.scheduler.ActorScheduler;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.test.util.asserts.grpc.ClientStatusExceptionAssert;
 import io.camunda.zeebe.test.util.socket.SocketUtil;
+import io.camunda.zeebe.util.micrometer.MicrometerUtil;
 import io.grpc.Status.Code;
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import io.netty.util.NetUtil;
 import java.io.IOException;
 import java.net.InetAddress;
@@ -46,13 +49,14 @@ import org.junit.jupiter.params.provider.MethodSource;
 
 @Execution(ExecutionMode.CONCURRENT)
 class UnavailableBrokersTest {
-  static Gateway gateway;
-  static AtomixCluster cluster;
-  static ActorScheduler actorScheduler;
-  static ZeebeClient client;
-  static BrokerClient brokerClient;
-  static JobStreamClient jobStreamClient;
-  static BrokerTopologyManagerImpl topologyManager;
+  private static final MeterRegistry meterRegistry = new SimpleMeterRegistry();
+  private static Gateway gateway;
+  private static AtomixCluster cluster;
+  private static ActorScheduler actorScheduler;
+  private static ZeebeClient client;
+  private static BrokerClient brokerClient;
+  private static JobStreamClient jobStreamClient;
+  private static BrokerTopologyManagerImpl topologyManager;
 
   @BeforeAll
   static void setUp() throws IOException {
@@ -68,7 +72,8 @@ class UnavailableBrokersTest {
 
     topologyManager =
         new BrokerTopologyManagerImpl(
-            () -> cluster.getMembershipService().getMembers(), BrokerClientTopologyMetrics.NOOP);
+            () -> cluster.getMembershipService().getMembers(),
+            new BrokerClientTopologyMetrics(meterRegistry));
     actorScheduler.submitActor(topologyManager).join();
     cluster.getMembershipService().addListener(topologyManager);
 
@@ -79,7 +84,7 @@ class UnavailableBrokersTest {
             cluster.getEventService(),
             actorScheduler,
             topologyManager,
-            BrokerClientRequestMetrics.NOOP);
+            new BrokerClientRequestMetrics(meterRegistry));
     jobStreamClient = new JobStreamClientImpl(actorScheduler, cluster.getCommunicationService());
     jobStreamClient.start().join();
 
@@ -104,7 +109,8 @@ class UnavailableBrokersTest {
         jobStreamClient,
         topologyManager,
         actorScheduler,
-        () -> cluster.stop().join());
+        () -> cluster.stop().join(),
+        () -> MicrometerUtil.closeRegistry(meterRegistry));
   }
 
   @ParameterizedTest(name = "{0}")
