+++ zeebe/gateway-grpc/src/test/java/io/camunda/zeebe/gateway/interceptors/InterceptorIT.java
@@ -75,7 +75,8 @@ final class InterceptorIT {
             .build();
     topologyManager =
         new BrokerTopologyManagerImpl(
-            () -> cluster.getMembershipService().getMembers(), BrokerClientTopologyMetrics.NOOP);
+            () -> cluster.getMembershipService().getMembers(),
+            new BrokerClientTopologyMetrics(new SimpleMeterRegistry()));
     cluster.getMembershipService().addListener(topologyManager);
 
     brokerClient =
@@ -85,7 +86,7 @@ final class InterceptorIT {
             cluster.getEventService(),
             scheduler,
             topologyManager,
-            BrokerClientRequestMetrics.NOOP);
+            new BrokerClientRequestMetrics(new SimpleMeterRegistry()));
 
     jobStreamClient = new JobStreamClientImpl(scheduler, cluster.getCommunicationService());
     gateway = new Gateway(config, brokerClient, scheduler, jobStreamClient.streamer());
